nginx 1.25.1 macos updates

+++ nginx/1.25.1/macos/install.sh
@@ -9,6 +9,10 @@ if [ ! -d "$HOME/programs" ]; then
 	mkdir "$HOME/programs"
 fi
 
+if [ ! -d "$HOME/logs" ]; then
+	mkdir "$HOME/logs"
+fi
+
 if [ ! -d "$HOME/sources/$FOLDER_NAME" ]; then
 	mkdir "$HOME/sources/$FOLDER_NAME"
 fi
@@ -17,37 +21,62 @@ if [ ! -d "$HOME/programs/$FOLDER_NAME" ]; then
 	mkdir "$HOME/programs/$FOLDER_NAME"
 fi
 
+if [ ! -d "$HOME/logs/$FOLDER_NAME" ]; then
+	mkdir "$HOME/logs/$FOLDER_NAME"
+fi
+
+if [ ! -d "$HOME/logs/$FOLDER_NAME/$VERSION" ]; then
+	mkdir "$HOME/logs/$FOLDER_NAME/$VERSION"
+fi
+
+if [ ! -e $HOME/workspace/myProjects/config-samples/$FOLDER_NAME/$VERSION/macos/nginx.conf ]; then
+	printf "nginx.conf not found\n"
+	exit
+fi
+
 if [ ! -d "$HOME/programs/$FOLDER_NAME/$VERSION" ]; then
 	mkdir "$HOME/programs/$FOLDER_NAME/$VERSION"
 
 	cd $HOME/sources/$FOLDER_NAME
 
-	wget "http://nginx.org/download/nginx-"$VERSION".tar.gz"
-	tar -xvf "nginx-"$VERSION".tar.gz"
+	printf "${bold}${yellow}Installing $FOLDER_NAME${clear}\n"
+
+	printf "\t${bold}${green}Downloading source code${clear}\n"
+	wget -q --show-progress "http://nginx.org/download/nginx-"$VERSION".tar.gz"
+	printf "\t${bold}${green}Extracting source code${clear}\n"
+	tar -xf "nginx-"$VERSION".tar.gz"
 	mv "nginx-"$VERSION $VERSION
 	cd $VERSION
-	./configure --prefix=$HOME/programs/nginx/$VERSION
-	make
-	sudo make install
+	printf "\t${bold}${green}Configuring${clear}\n"
+	./configure --help > $HOME/logs/$FOLDER_NAME/$VERSION/configureHelp.txt 2>&1
+	./configure --prefix=$HOME/programs/nginx/$VERSION > $HOME/logs/$FOLDER_NAME/$VERSION/configureOutput.txt 2>&1
+	printf "\t${bold}${green}Making${clear}\n"
+	make > $HOME/logs/$FOLDER_NAME/$VERSION/makeOutput.txt 2>&1
+	printf "\t${bold}${green}Installing${clear}\n"
+	echo $USER_PASSWORD | sudo -S make install > $HOME/logs/$FOLDER_NAME/$VERSION/installOutput.txt 2>&1
 
-	cd $HOME/programs/$FOLDER_NAME/$VERSION
-	sudo chown -R $(whoami) .
+	if [ -e "$HOME/programs/$FOLDER_NAME/$VERSION/sbin/nginx" ]; then
+		cd $HOME/programs/$FOLDER_NAME/$VERSION
+		echo $USER_PASSWORD | sudo -S chown -R $(whoami) .
 
-	touch .envrc
-	echo 'export PATH=$HOME/programs/'"$FOLDER_NAME/$VERSION/sbin:"'$PATH' >> .envrc
-	echo "" >> .envrc
-	direnv allow
+		touch .envrc
+		echo 'export PATH=$HOME/programs/'"$FOLDER_NAME/$VERSION/sbin:"'$PATH' >> .envrc
+		echo "" >> .envrc
+		direnv allow
 
-	rm conf/nginx.conf
-	ln -s ~/workspace/myProjects/config-samples/$FOLDER_NAME/$VERSION/macos/nginx.conf conf/nginx.conf
+		mv conf/nginx.conf ~/workspace/myProjects/config-samples/$FOLDER_NAME/$VERSION/macos/nginx.conf.default
+		ln -s ~/workspace/myProjects/config-samples/$FOLDER_NAME/$VERSION/macos/nginx.conf conf/nginx.conf
+		echo "<html><body><h1>It works! (version: $VERSION)</h1></body></html>" > html/index.html
 
-	touch start.sh
-	echo "nginx" >> start.sh
+		touch start.sh
+		echo "nginx" >> start.sh
 
-	touch stop.sh
-	echo 'nginx -s stop' >> stop.sh
+		touch stop.sh
+		echo 'nginx -s stop' >> stop.sh
 
-	cd $HOME/sources/$FOLDER_NAME
-	rm -rf $VERSION
-	rm "nginx-"$VERSION".tar.gz"
+		printf "\t${bold}${green}Clearing${clear}\n"
+		cd $HOME/sources/$FOLDER_NAME
+		rm -rf $VERSION
+		rm "nginx-"$VERSION".tar.gz"
+	fi
 fi
